Remove excess `supabase` declaration
