+++ src/app/api/todo-lists/create-from-template/route.ts
@@ -1,8 +1,8 @@
 import { withRevalidation } from '@/lib/api/withRevalidation';
-import { supabase } from '@/lib/supabase';
-import { isValidTemplateId } from '@/types/todo-list-template';
+import { supabase } from '@/lib/db';
 import { TODO_TEMPLATES } from '@/lib/templates/todo-templates';
 import { THEME_COLORS } from '@/types/todo-list';
+import { isValidTemplateId } from '@/types/todo-list-template';
 import { NextRequest, NextResponse } from 'next/server';
 
 const getRandomTheme = () => {
