+++ src/app/api/todo-lists/route.ts
@@ -1,5 +1,5 @@
 import { withRevalidation } from '@/lib/api/withRevalidation';
-import { supabase } from '@/lib/supabase';
+import { supabase } from '@/lib/db';
 import { NextRequest, NextResponse } from 'next/server';
 
 export async function GET(request: NextRequest) {
